backend/requirements.txt
fastapi[standard]
sqlalchemy
requests
beautifulsoup4
openai
anthropic
python-dotenv
playwright
httpx
imageio
imageio-ffmpeg
